Fix Win64 build by adding a static_cast.
BUG=
TEST=

Review URL: https://chromiumcodereview.appspot.com/10784040

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@12113 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/hydrogen.h
@@ -872,7 +872,7 @@ class HGraphBuilder: public AstVisitor {
   void VisitDeclarations(ZoneList<Declaration*>* declarations);
 
   void* operator new(size_t size, Zone* zone) {
-    return zone->New(size);
+    return zone->New(static_cast<int>(size));
   }
   void operator delete(void* ptr) { }
